+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## 2024-05-05
 
+### Changed
+
+- Perks now display their type on the item cards and item info dialogs.
+
 ### Fixed
 
 - Fixed all discrepancies with Mutator item descriptions and stats from the wiki.